added auth middleware (protext api routes) and auth header.

+++ frontend/src/components/Workouts.jsx
@@ -1,14 +1,20 @@
 import { useWorkoutContext } from '../hooks/useWorkoutContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // date-format using date-fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const Workouts = ({ workout: { title, load, reps, _id, createdAt } }) => {
   const { dispatch } = useWorkoutContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
     const res = await fetch(`/api/workouts/${_id}`, {
       method: 'DELETE',
+      Authorization: `Bearer ${user.token}`,
     })
 
     const json = await res.json()
